instance builder change, add skip window
skip window to be used for re-calculating after a failed calculation for a line, to allow the user to go back and change value of data in the database results from searching using that NAICS code, or just change uncertainty limits for that process (i.e. skip changing the data from the database).